Update README.md
Hej, Hello - I suggest to "hardcode" the units in Precipitation Total in order to guide users in their setup of the Items and prevent them from receiving by default data in wrong units. 

rason:
It seems the use of the standard declaration (%unit% ) does not work correctly, and renders the unit for Precipation Total in SI units to the quite scale of meters "m" 
As a workaround for the Precipation rate was found.; https://community.openhab.org/t/weathercompany-binding-weatherunderground-replacement-for-pws-data-uploaders-only/69122/114?u=mdnx  and it seems, that the underlying technical problem persists unfortunately.